Move static subsettingsbuilder to helper class
