+++ plugin/src/main/java/net/codingarea/challenges/plugin/challenges/custom/settings/action/impl/ClearInventoryAction.java
@@ -2,6 +2,7 @@ package net.codingarea.challenges.plugin.challenges.custom.settings.action.impl;
 
 import java.util.Map;
 import net.codingarea.challenges.plugin.challenges.custom.settings.action.AbstractChallengePlayerTargetAction;
+import net.codingarea.challenges.plugin.challenges.type.helper.SubSettingsHelper;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 
@@ -12,7 +13,7 @@ import org.bukkit.entity.Player;
 public class ClearInventoryAction extends AbstractChallengePlayerTargetAction {
 
   public ClearInventoryAction(String name) {
-    super(name, createEntityTargetSettingsBuilder(false, true));
+    super(name, SubSettingsHelper.createEntityTargetSettingsBuilder(false, true));
   }
 
   @Override
