+++ plugin/src/main/java/net/codingarea/challenges/plugin/challenges/custom/settings/action/impl/DropRandomItemAction.java
@@ -3,6 +3,7 @@ package net.codingarea.challenges.plugin.challenges.custom.settings.action.impl;
 import java.util.Map;
 import net.codingarea.challenges.plugin.challenges.custom.settings.action.AbstractChallengePlayerTargetAction;
 import net.codingarea.challenges.plugin.challenges.implementation.challenge.RandomItemDroppingChallenge;
+import net.codingarea.challenges.plugin.challenges.type.helper.SubSettingsHelper;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 
@@ -13,7 +14,7 @@ import org.bukkit.entity.Player;
 public class DropRandomItemAction extends AbstractChallengePlayerTargetAction {
 
   public DropRandomItemAction(String name) {
-    super(name, createEntityTargetSettingsBuilder(false, true));
+    super(name, SubSettingsHelper.createEntityTargetSettingsBuilder(false, true));
   }
 
   @Override
