+++ plugin/src/main/java/net/codingarea/challenges/plugin/challenges/custom/settings/action/impl/KillEntityAction.java
@@ -3,6 +3,7 @@ package net.codingarea.challenges.plugin.challenges.custom.settings.action.impl;
 import java.util.Map;
 import net.codingarea.challenges.plugin.challenges.custom.settings.action.AbstractChallengeTargetAction;
 import net.codingarea.challenges.plugin.challenges.type.helper.ChallengeHelper;
+import net.codingarea.challenges.plugin.challenges.type.helper.SubSettingsHelper;
 import org.bukkit.Material;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.LivingEntity;
@@ -15,7 +16,7 @@ import org.bukkit.entity.Player;
 public class KillEntityAction extends AbstractChallengeTargetAction {
 
   public KillEntityAction(String name) {
-    super(name, createEntityTargetSettingsBuilder(true));
+    super(name, SubSettingsHelper.createEntityTargetSettingsBuilder(true));
   }
 
   @Override
