+++ plugin/src/main/java/net/codingarea/challenges/plugin/challenges/custom/settings/trigger/AbstractChallengeTrigger.java
@@ -2,17 +2,11 @@ package net.codingarea.challenges.plugin.challenges.custom.settings.trigger;
 
 import java.util.LinkedHashMap;
 import java.util.function.Supplier;
-import net.anweisen.utilities.bukkit.utils.misc.BukkitReflectionUtils;
-import net.anweisen.utilities.common.misc.StringUtils;
 import net.codingarea.challenges.plugin.Challenges;
 import net.codingarea.challenges.plugin.challenges.custom.settings.AbstractChallengeSetting;
 import net.codingarea.challenges.plugin.challenges.custom.settings.sub.SubSettingsBuilder;
-import net.codingarea.challenges.plugin.challenges.custom.settings.sub.builder.ChooseMultipleItemSubSettingBuilder;
 import net.codingarea.challenges.plugin.content.Message;
-import net.codingarea.challenges.plugin.utils.item.DefaultItem;
 import net.codingarea.challenges.plugin.utils.item.ItemBuilder;
-import org.bukkit.Material;
-import org.bukkit.entity.EntityType;
 import org.bukkit.inventory.ItemStack;
 
 /**
@@ -21,13 +15,6 @@ import org.bukkit.inventory.ItemStack;
  */
 public abstract class AbstractChallengeTrigger extends AbstractChallengeSetting implements
     IChallengeTrigger {
-
-  public static final String ENTITY_TYPE = "entity_type",
-        BLOCK = "block",
-        ANY = "any",
-        ITEM = "item",
-        LIQUID = "liquid",
-        TARGET_ENTITY = "target_entity";
 
   public AbstractChallengeTrigger(String name,
       SubSettingsBuilder subSettingsBuilder) {
@@ -55,35 +42,6 @@ public abstract class AbstractChallengeTrigger extends AbstractChallengeSetting 
     }
 
     return map;
-  }
-
-  public static ChooseMultipleItemSubSettingBuilder createEntityTypeSettingsBuilder() {
-    return SubSettingsBuilder.createChooseMultipleItem(AbstractChallengeTrigger.ENTITY_TYPE).fill(builder -> {
-      builder.addSetting(AbstractChallengeTrigger.ANY, new ItemBuilder(Material.NETHER_STAR, Message.forName("item-custom-trigger-entity_type-any")).build());
-      builder.addSetting("PLAYER", new ItemBuilder(Material.PLAYER_HEAD, Message.forName("item-custom-trigger-entity_type-player")).build());
-      for (EntityType type : EntityType.values()) {
-        if (!type.isSpawnable() || !type.isAlive()) continue;
-        try {
-          Material spawnEgg = Material.valueOf(type.name() + "_SPAWN_EGG");
-          builder.addSetting(type.name(), new ItemBuilder(spawnEgg,
-              DefaultItem.getItemPrefix() + StringUtils.getEnumName(type)).build());
-        } catch (Exception ex) {
-          builder.addSetting(type.name(), new ItemBuilder(Material.STRUCTURE_VOID,
-              DefaultItem.getItemPrefix() + StringUtils.getEnumName(type)));
-        }
-      }
-    });
-  }
-
-  public static ChooseMultipleItemSubSettingBuilder createBlockSettingsBuilder() {
-    return SubSettingsBuilder.createChooseMultipleItem(AbstractChallengeTrigger.BLOCK).fill(builder -> {
-      builder.addSetting(AbstractChallengeTrigger.ANY, new ItemBuilder(Material.NETHER_STAR, Message.forName("item-custom-trigger-block-any")).build());
-      for (Material material : Material.values()) {
-        if (material.isBlock() && material.isItem() && !BukkitReflectionUtils.isAir(material)) {
-          builder.addSetting(material.name(), new ItemBuilder(material, DefaultItem.getItemPrefix() + StringUtils.getEnumName(material)).build());
-        }
-      }
-    });
   }
 
 }
