+++ plugin/src/main/java/net/codingarea/challenges/plugin/challenges/custom/settings/trigger/impl/BreakBlockTrigger.java
@@ -1,6 +1,7 @@
 package net.codingarea.challenges.plugin.challenges.custom.settings.trigger.impl;
 
 import net.codingarea.challenges.plugin.challenges.custom.settings.trigger.AbstractChallengeTrigger;
+import net.codingarea.challenges.plugin.challenges.type.helper.SubSettingsHelper;
 import org.bukkit.Material;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
@@ -13,7 +14,7 @@ import org.bukkit.event.block.BlockBreakEvent;
 public class BreakBlockTrigger extends AbstractChallengeTrigger {
 
   public BreakBlockTrigger(String name) {
-    super(name, createBlockSettingsBuilder());
+    super(name, SubSettingsHelper.createBlockSettingsBuilder());
   }
 
   @Override
